+++ laundry-ops/android/app/src/main/java/com/tiftazani/laundryops/CuciinApp.kt
@@ -1,5 +1,11 @@
 package com.tiftazani.laundryops
 
 import android.app.Application
+import com.tiftazani.laundryops.data.FirebaseCloud
 
-class CuciinApp : Application()
+class CuciinApp : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        FirebaseCloud.init(this)
+    }
+}